Update README, enhance project structure, and implement HTTP server features with multithreading support
