+++ include/common/cli_style.h
@@ -29,4 +29,4 @@
 #define BG_CYAN "\033[46m"
 #define BG_WHITE "\033[47m"
 
-#endif // CLI_STYLE_H
+#endif /* CLI_STYLE_H */
