+++ include/server/config.h
@@ -1,0 +1,10 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+/**
+ * Server configuration
+ */
+#define PORT 8080
+#define BUFFER_SIZE 1024
+
+#endif /* CONFIG_H */
