--- includes/index.h
@@ -1,6 +1,0 @@
-#ifndef INDEX
-#define INDEX
-
-#include "./common/lib/cli_style.h"
-
-#endif
