--- includes/socket_setup.h
@@ -1,6 +1,0 @@
-#ifndef TCP_SOCKET_H
-#define TCP_SOCKET_H
-
-int socket_tcp();
-
-#endif
