+++ www/index.html
@@ -4,11 +4,63 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
+    <title>Marouane El Moujahid - Introduction</title>
+    <style>
+        body {
+            font-family: Arial, sans-serif;
+            line-height: 1.6;
+            margin: 0;
+            padding: 20px;
+            max-width: 800px;
+            margin: 0 auto;
+        }
+
+        header {
+            text-align: center;
+            margin-bottom: 30px;
+        }
+
+        h1 {
+            color: #333;
+        }
+
+        .section {
+            margin-bottom: 25px;
+        }
+    </style>
 </head>
 
 <body>
+    <header>
+        <h1>Marouane El Moujahid</h1>
+        <p>Welcome to my personal page</p>
+    </header>
 
+    <div class="section">
+        <h2>About Me</h2>
+        <p>Hello! I'm Marouane El Moujahid. This is my personal introduction page where you can learn more about me, my
+            skills, and interests.</p>
+    </div>
+
+    <div class="section">
+        <h2>Skills</h2>
+        <ul>
+            <li>Programming</li>
+            <li>Web Development</li>
+            <li>Problem Solving</li>
+            <!-- Add more skills as needed -->
+        </ul>
+    </div>
+
+    <div class="section">
+        <h2>Contact</h2>
+        <p>Feel free to reach out to me via email: <a href="mailto:your.email@example.com">your.email@example.com</a>
+        </p>
+    </div>
+
+    <footer>
+        <p>&copy; 2023 Marouane El Moujahid</p>
+    </footer>
 </body>
 
 </html>
